Implement central zonotopal D-space basis
Basis implemented under the assumption that the derivative of a D-space
basis polynomial can be represented in terms of D-space basis
polynomials which lie below it in the external order.  Initial testing
of the algorithm on modestly large examples bears this conjecture out.

Additionally modified the output of the basis algorithms to be a Python
dictionary whose keys are frozensets for the bases of the matroid, and
whose values are the corresponding polynomials.  Will need to update
any connected function calls to reflect this change.