Add workaround for Sage linear algebra bug
Sage 9.4 has an issue related to a backend library which does the linear
algebra computations for class FreeModule, so that several methods used
to solve systems of linear equations over the rational field don't work.
The following trac ticket most closely matches the observed symptoms:

https://trac.sagemath.org/ticket/32447

This broke the method used to find a relative normal vector of a
subspace within another subspace, so as a workaround this commit does
orthogonal projection computations by hand.  According to the ticket
notes, the issue should be resolved in Sage 9.5.